backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-multipart
groq
pydantic
fpdf2
PyJWT
cryptography
python-dotenv
requests
passlib
bcrypt==3.1.7
python-jose[cryptography]
